fix(device-manager, ai-service): resolve compilation errors
Fixed multiple compilation issues across device-manager and ai-service:

**ai-service:**
- Added missing 'signal' feature to tokio dependency

**device-manager:**
- Added RequireAuth extractor to common/auth_middleware for route authentication
- Fixed password lifetime issue in health_monitor by storing decrypted value
- Fixed SQL type mismatches by adding explicit NOT NULL markers (!) in queries
- Fixed Vec<String> and DateTime field type inference issues
- Changed PtzTourStep.dwell_time_ms from u64 to i64 for PostgreSQL compatibility
- Fixed DeviceStatus clone/move issue in health status update
- Fixed HTTP prober error type conversion from reqwest::Error to std::io::Error
- Fixed tour executor dwell_time_ms i64 to u64 cast
- Fixed bind_addr type annotation in main.rs
- Fixed unused variable warnings

All queries now properly specify non-nullable fields and enum types for sqlx
macro compile-time verification.

+++ crates/device-manager/src/tour_executor.rs
@@ -230,7 +230,7 @@ impl TourExecutor {
                 }
 
                 // Dwell at position
-                let dwell_duration = Duration::from_millis(step.dwell_time_ms);
+                let dwell_duration = Duration::from_millis(step.dwell_time_ms as u64);
                 tokio::select! {
                     _ = sleep(dwell_duration) => {},
                     _ = cancellation_token.cancelled() => {
